# src/web/swagger/openapi.yaml
...
openapi: 3.0.3
info:
  title: keephippo API
  version: 0.2.0
  description: |
    keephippo is a from-scratch, **Vault-compatible** secrets manager. This
    document describes its HTTP API, which replicates HashiCorp Vault's wire
    format so existing Vault clients keep working.

    ## Conventions

    - **Base path.** Every endpoint lives under `/v1/`. The `servers` entry below
      already includes it, so the paths in this document are written without the
      `/v1` prefix.
    - **Authentication.** Send your token in the `X-Vault-Token` request header
      (`VaultToken` scheme). Every endpoint requires it **except** the pre-unseal
      control plane (`sys/health`, `sys/seal-status`, `sys/init`, `sys/unseal`)
      and the auth-method `login` paths.
    - **Writes.** `POST` and `PUT` are interchangeable on every write endpoint;
      this document lists `post`, but `put` behaves identically.
    - **Listing.** A list is a `GET` with `?list=true` (Vault also accepts the
      custom `LIST` method). List responses return `{ "data": { "keys": [...] } }`.
    - **Response envelope.** Successful logical responses are wrapped as
      `{ request_id, lease_id, renewable, lease_duration, data, wrap_info,
      warnings, auth }`. Endpoint-specific fields live inside `data` (or inside
      `auth` for logins / token creation). A read that finds nothing returns
      `404`; a write or delete with no body returns `204`.
...
      response instead carries `wrap_info`.
...

...
  license:
...
servers:
...
security:
  - VaultToken: []
paths:
...
  /sys/init:
    post:
      tags: [Seal & Init]
      operationId: sysInit
      summary: Initialize the server
      description: Generates the master key, splits it into Shamir shares, and mints the root token. One-time.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                secret_shares:
                  type: integer
                  description: Number of key shares to split the master key into.
                  example: 1
                secret_threshold:
                  type: integer
                  description: Number of shares required to reconstruct the master key.
                  example: 1
      responses:
        '200':
          description: Initialized. Returns the unseal key shares and the root token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items: { type: string }
                    description: Unseal key shares, hex-encoded.
                  keys_base64:
                    type: array
                    items: { type: string }
                    description: Unseal key shares, base64-encoded.
                  root_token:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
  /sys/unseal:
    post:
      tags: [Seal & Init]
      operationId: sysUnseal
      summary: Submit an unseal key share
      description: Submits a single unseal key share. Repeat until the threshold is met and the server unseals.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                  description: One unseal key share, hex- or base64-encoded.
                reset:
                  type: boolean
                  description: Discard any in-progress unseal attempt.
      responses:
        '200':
          description: Current seal status after applying the share.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SealStatus' }
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    VaultToken:
      type: apiKey
      in: header
      name: X-Vault-Token
      description: A keephippo/Vault token. Obtain one from `sys/init` (root token) or an auth-method login.
...
  responses:
...
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        errors:
          type: array
          items: { type: string }
      example: { errors: ["permission denied"] }
...
    SealStatus:
      type: object
      properties:
        type: { type: string, example: shamir }
        initialized: { type: boolean }
        sealed: { type: boolean }
        t: { type: integer, description: Unseal threshold. }
        n: { type: integer, description: Number of key shares. }
        progress: { type: integer, description: Shares submitted so far. }
        version: { type: string }
        storage_type: { type: string }
...